@@ -13,19 +13,19 @@
 			"role": "Lead developer"
 	}],
 	"require": {
-		"php": ">=5.6",
+		"php": ">=7.0",
 		"squizlabs/php_codesniffer": "^3.3.0",
-		"wptrt/wpthemereview": "^0.1.0"
+		"wptrt/wpthemereview": "^0.1.0",
+		"dealerdirect/phpcodesniffer-composer-installer": "^0.5.0"
 	},
 	"require-dev": {
 		"phpcompatibility/php-compatibility": "^9.0",
-		"roave/security-advisories"         : "dev-master",
-		"dealerdirect/phpcodesniffer-composer-installer": "^0.5.0"
+		"roave/security-advisories"         : "dev-master"
 	},
 	"autoload": {
 		"classmap": [
-			"admin/",
-			"includes/"
+			"src/",
+			"views/"
 		]
 	},
 	"scripts": {
