@@ -1,60 +1,60 @@
 <?php
 /**
- * File that holds helper class with static helper methods
+ * The helpers trait file
  *
- * @since 0.2.0
- *
- * @package Theme_Sniffer\Admin
+ * @since   2.0.0
+ * @package Theme_Sniffer\Helpers
  */
 
-namespace Theme_Sniffer\Admin;
+declare( strict_types=1 );
+
+namespace Theme_Sniffer\Helpers;
 
 /**
- * Class that holds the helper methods.
+ * Sniffer helpers trait
  *
- * @package Theme_Sniffer\Admin
+ * This trait contains some helper methods
  */
-class Helpers {
-
+trait Sniffer_Helpers {
 	/**
 	 * Returns standards
 	 *
 	 * Includes a 'theme_sniffer_add_standards' filter, so that user can add their own standard. The standard has to be added
 	 * in the composer before bundling the plugin.
 	 *
-	 * @since 0.2.0 Added filter so that user can add their own standards.
+	 * @since 0.2.0 Added filter so that user can add their own standards. Moved to a trait.
 	 * @since 0.1.3
 	 *
 	 * @return array Standards details.
 	 */
-	public static function get_wpcs_standards() {
-		$standards = array(
-			'wordpress-theme' => array(
+	public function get_wpcs_standards() : array {
+		$standards = [
+			'wordpress-theme' => [
 				'label'       => 'WPThemeReview',
 				'description' => esc_html__( 'Ruleset for WordPress theme review requirements (Required)', 'theme-sniffer' ),
 				'default'     => 1,
-			),
-			'wordpress-core'  => array(
+			],
+			'wordpress-core'  => [
 				'label'       => 'WordPress-Core',
 				'description' => esc_html__( 'Main ruleset for WordPress core coding standards (Optional)', 'theme-sniffer' ),
 				'default'     => 0,
-			),
-			'wordpress-extra' => array(
+			],
+			'wordpress-extra' => [
 				'label'       => 'WordPress-Extra',
 				'description' => esc_html__( 'Extended ruleset for recommended best practices (Optional)', 'theme-sniffer' ),
 				'default'     => 0,
-			),
-			'wordpress-docs'  => array(
+			],
+			'wordpress-docs'  => [
 				'label'       => 'WordPress-Docs',
 				'description' => esc_html__( 'Additional ruleset for WordPress inline documentation standards (Optional)', 'theme-sniffer' ),
 				'default'     => 0,
-			),
-			'wordpress-vip'   => array(
+			],
+			'wordpress-vip'   => [
 				'label'       => 'WordPress-VIP',
 				'description' => esc_html__( 'Extended ruleset for WordPress VIP coding requirements (Optional)', 'theme-sniffer' ),
 				'default'     => 0,
-			),
-		);
+			],
+		];
 
 		if ( has_filter( 'theme_sniffer_add_standards' ) ) {
 			$standards = apply_filters( 'theme_sniffer_add_standards', $standards );
@@ -66,12 +66,12 @@ class Helpers {
 	/**
 	 * Return all the active themes
 	 *
-	 * @since  0.2.0
+	 * @since  0.2.0 Moved to a trait.
 	 * @return array Array of active themes.
 	 */
-	public static function get_active_themes() {
+	public function get_active_themes() : array {
 		$all_themes = wp_get_themes();
-		$themes     = array();
+		$themes     = [];
 
 		if ( ! empty( $all_themes ) ) {
 			foreach ( $all_themes as $key => $theme ) {
@@ -85,14 +85,13 @@ class Helpers {
 	/**
 	 * Returns PHP versions.
 	 *
-	 * @since 0.2.0 Added PHP 7.x versions
+	 * @since 0.2.0 Added PHP 7.x versions. Moved to a trait.
 	 * @since 0.1.3
 	 *
 	 * @return array PHP versions.
 	 */
-	public static function get_php_versions() {
-
-		$output = array(
+	public function get_php_versions() : array {
+		return [
 			'5.2',
 			'5.3',
 			'5.4',
@@ -102,21 +101,20 @@ class Helpers {
 			'7.1',
 			'7.2',
 			'7.3',
-		);
-
-		return $output;
+		];
 	}
 
 	/**
 	 * Returns theme tags.
 	 *
+	 * @since 0.2.0 Moved to a trait.
 	 * @since 0.1.3
 	 *
-	 * @return array PHP versions.
+	 * @return array Theme tags array.
 	 */
-	public static function get_theme_tags() {
+	public function get_theme_tags() : array {
 
-		$tags['allowed_tags'] = array(
+		$tags['allowed_tags'] = [
 			'two-columns',
 			'three-columns',
 			'four-columns',
@@ -144,9 +142,9 @@ class Helpers {
 			'theme-options',
 			'threaded-comments',
 			'translation-ready',
-		);
+		];
 
-		$tags['subject_tags'] = array(
+		$tags['subject_tags'] = [
 			'blog',
 			'e-commerce',
 			'education',
@@ -156,8 +154,18 @@ class Helpers {
 			'news',
 			'photography',
 			'portfolio',
-		);
+		];
 
 		return $tags;
+	}
+
+	/**
+	 * Helper method that returns the default stnadard
+	 *
+	 * @since 0.2.0
+	 * @return string Name of the default standard
+	 */
+	public function get_default_standard() : string {
+		return 'WPThemeReview';
 	}
 }
